LayoutTests/media/media-source/media-source-seek-back-after-ended-expected.txt

This tests that a SourceBuffer can accept an initialization segment and a media segment and fire "update" events, the element can play and seek back properly when it's ended.

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(source.duration = loader.duration())
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(source.endOfStream())
RUN(await video.play())
EXPECTED (video.paused == 'false') OK
EVENT(ended)
RUN(video.currentTime = 0)
EVENT(seeked)
RUN(await video.play())
EXPECTED (video.paused == 'false') OK
EXPECTED (video.currentTime > '1') OK
END OF TEST
